factored hotkeys out of screen into their own container component
